requirements.txt
fastapi
uvicorn
boto3
python-multipart
python-decouple
pytest
httpx
